test: Reproduce expected incorrect results for large output sizes
`FoldNd` unfolds an index tensor which is represented as `float32`.
These indices are then converted to `long` and used to `scatter_add` the
input entries into the result tensor. This will not work if the indices
cannot be represented as `float32`. This happens if the output size
exceeds a size is 2 ** 24 = 116777217 (see
https://stackoverflow.com/q/27207149p for details).