[FIX] Treat padding explicitly when scattering, then unpad
This should fix #30.

The problem was in the creation of the tensor that maps the patch values
back to the positions in the original image. The code was using `unfold`
with non-zero padding, which would lead to a lot of patch values being
assigned to location 0. The fix is to deal with padding explicitly so
that the `unfold` operation uses `padding=0`. Then the index tensor is
correct, and after scattering we have to un-pad the image to obtain the
output.